refactor(extensions): give the manifest version one definition
The constant was declared in both manifest.ts and schemas.ts with the
same value. Two declarations of one fact invite being raised in one place
and left behind in the other, at which point the newer-format check and
the install-manifest check disagree about what a version means.

Neither module could own it. The schemas define the format but load
behind a dynamic import, so that a validation library never reaches the
startup path; discovery asks whether a manifest is newer while building a
listing and needs the answer synchronously. So the version and the two
predicates over it now live in a module of their own, which imports
nothing at runtime, and both sides read it from there.
