+++ src/lib/extensions/manifest.test.ts
@@ -2,11 +2,9 @@ import { chmod, mkdtemp, readFile, rm, writeFile } from 'node:fs/promises'
 import { tmpdir } from 'node:os'
 import { join } from 'node:path'
 import { afterEach, beforeEach, describe, expect, it } from 'vitest'
+import { isFromNewerFormat, MANIFEST_VERSION, manifestVersionOf } from './manifest-format.js'
 import {
     findManifestProblems,
-    isFromNewerFormat,
-    MANIFEST_VERSION,
-    manifestVersionOf,
     readAuthoredManifest,
     readInstalledManifest,
     writeInstalledManifest,
